refactor: Moving transaction pool error type from `fuel-core-types` to `fuel-core-txpool`
