+++ crates/services/txpool/src/txpool.rs
@@ -149,7 +149,7 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
             for remove in removed.iter() {
                 self.remove_tx(&remove.id());
             }
-            return removed
+            return removed;
         }
         Vec::new()
     }
@@ -245,7 +245,7 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
             tokio::time::Instant::now().checked_sub(self.config.transaction_ttl)
         else {
             // TTL is so big that we don't need to prune any transactions
-            return vec![]
+            return vec![];
         };
 
         let mut result = vec![];
@@ -256,7 +256,7 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
                 let removed = self.remove_inner(&oldest_tx);
                 result.extend(removed.into_iter());
             } else {
-                break
+                break;
             }
         }
 
@@ -269,10 +269,10 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
                 Input::CoinSigned(CoinSigned { utxo_id, owner, .. })
                 | Input::CoinPredicate(CoinPredicate { utxo_id, owner, .. }) => {
                     if self.config.blacklist.contains_coin(utxo_id) {
-                        return Err(Error::BlacklistedUTXO(*utxo_id))
+                        return Err(Error::BlacklistedUTXO(*utxo_id));
                     }
                     if self.config.blacklist.contains_address(owner) {
-                        return Err(Error::BlacklistedOwner(*owner))
+                        return Err(Error::BlacklistedOwner(*owner));
                     }
                 }
                 Input::Contract(contract) => {
@@ -281,7 +281,7 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
                         .blacklist
                         .contains_contract(&contract.contract_id)
                     {
-                        return Err(Error::BlacklistedContract(contract.contract_id))
+                        return Err(Error::BlacklistedContract(contract.contract_id));
                     }
                 }
                 Input::MessageCoinSigned(MessageCoinSigned {
@@ -309,13 +309,13 @@ impl<ViewProvider, WasmChecker> TxPool<ViewProvider, WasmChecker> {
                     ..
                 }) => {
                     if self.config.blacklist.contains_message(nonce) {
-                        return Err(Error::BlacklistedMessage(*nonce))
+                        return Err(Error::BlacklistedMessage(*nonce));
                     }
                     if self.config.blacklist.contains_address(sender) {
-                        return Err(Error::BlacklistedOwner(*sender))
+                        return Err(Error::BlacklistedOwner(*sender));
                     }
                     if self.config.blacklist.contains_address(recipient) {
-                        return Err(Error::BlacklistedOwner(*recipient))
+                        return Err(Error::BlacklistedOwner(*recipient));
                     }
                 }
             }
@@ -362,11 +362,11 @@ where
         self.check_blacklisting(tx.as_ref())?;
 
         if !tx.is_computed() {
-            return Err(Error::NoMetadata)
+            return Err(Error::NoMetadata);
         }
 
         if self.by_hash.contains_key(&tx.id()) {
-            return Err(Error::NotInsertedTxKnown)
+            return Err(Error::NotInsertedTxKnown);
         }
 
         // If upgrading transition function, at least check that it is valid wasm
@@ -392,7 +392,7 @@ where
             // limit is hit, check if we can push out lowest priced tx
             let lowest_ratio = self.by_ratio_gas_tip.lowest_value().unwrap_or_default();
             if lowest_ratio >= Ratio::new(tx.tip(), tx.max_gas()) {
-                return Err(Error::NotInsertedLimitHit)
+                return Err(Error::NotInsertedLimitHit);
             }
         }
         if self.config.metrics {
@@ -525,7 +525,7 @@ where
     M: Memory + Send + Sync + 'static,
 {
     if tx.is_mint() {
-        return Err(Error::NotSupportedTransactionType)
+        return Err(Error::NotSupportedTransactionType);
     }
 
     let tx: Checked<Transaction> = if utxo_validation {
